move forkchoice to config module and add comments
